+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.3.0
 - :bulb: Added `NumEmailsSent` score criteria
-- :hammer: Fixed `ZazoActivity` score criteria
 - :bulb: Added update info button to `contact#show` page
 - :bulb: Added links to other services to `contact#show` page
+- :hammer: Fixed `ZazoActivity` score criteria
+- :hammer: Fixed score calculation (additions caching)